refactor: reduce code by moving type SearchResult to core
